requirements.txt
Django==3.2.5
requests==2.28.2
djangorestframework==3.12.4
jsonpath-ng==1.5.3